% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hireLetter-function.R
\name{hireLetter}
\alias{hireLetter}
\title{hireLetter}
\usage{
hireLetter(data)
}
\arguments{
\item{data}{user inputs their data frame of applicant data}
}
\description{
function outputs shiny app consisting filtered data to view and download as a csv file of individuals who pass or fail their age and experience qualifications as well as draft acceptance and rejection letters
}
